+++ src/app/(admin)/admin/[user_id]/heizkostenabrechnung/localauswahl/weitermachen/[doc_id]/[path_slug]/dokumentenmanagement/page.tsx
@@ -23,7 +23,7 @@ export default async function DokumentenmanagementPage({
         <div className="py-6 px-9 max-medium:px-4 max-medium:py-4 h-[calc(100dvh-77px)] max-h-[calc(100dvh-77px)] max-xl:h-[calc(100dvh-53px)] max-xl:max-h-[calc(100dvh-53px)] max-medium:h-auto max-medium:max-h-none max-medium:overflow-y-auto grid grid-rows-[auto_1fr]">
             <Breadcrumb
                 backTitle="Abrechnung"
-                link={`${ROUTE_ADMIN}/${user_id}${ROUTE_HEIZKOSTENABRECHNUNG}/localauswahl/wietermachen/${doc_id}/abrechnungszeitraum`}
+                link={`${ROUTE_ADMIN}/${user_id}${ROUTE_HEIZKOSTENABRECHNUNG}/localauswahl/weitermachen/${doc_id}/abrechnungszeitraum`}
                 title={`Dokumentenmanagement`}
                 subtitle="Ladden Sie alle Ihre Rechnungen und Zahlungsnachweise die für die Betriebskostenabrechnung wichtig sind hoch."
             />
