+++ src/app/(admin)/admin/[user_id]/heizkostenabrechnung/objektauswahl/[objekt_id]/[doc_id]/[path_slug]/umlageschlussel/page.tsx
@@ -28,7 +28,7 @@ export default async function UmlageschlüsselPage({
     <div className="py-6 px-9 max-medium:px-4 max-medium:py-4 h-[calc(100dvh-77px)] max-h-[calc(100dvh-77px)] max-xl:h-[calc(100dvh-53px)] max-xl:max-h-[calc(100dvh-53px)] max-medium:h-auto max-medium:max-h-none max-medium:overflow-y-auto grid grid-rows-[auto_1fr]">
       <Breadcrumb
         backTitle="Abrechnung"
-        link={`${ROUTE_ADMIN}/${user_id}${ROUTE_HEIZKOSTENABRECHNUNG}/objektauswahl/${objekt_id}/${doc_id}/${path_slug}/gesamtkostens`}
+        link={`${ROUTE_ADMIN}/${user_id}${ROUTE_HEIZKOSTENABRECHNUNG}/objektauswahl/${objekt_id}/${doc_id}/${path_slug}/gesamtkosten`}
         title={`Umlageschlüssel erstellen`}
         subtitle="Bitte ergänzen Sie die Umlageschlüssel der Kostenkategorien. Mit dem Umlageschlüssel geben Sie als Vermieter vor, wie Sie die Kosten auf die verschiedenen Mietparteien im Haus verteilst."
       />
